+++ include/iota/api/extended.hpp
@@ -52,7 +52,7 @@ namespace API {
  */
 class Extended : public Core {
 public:
-  Extended(const std::string& host, const unsigned int& port,
+  Extended(const std::string& host, const uint16_t& port,
            Crypto::SpongeType cryptoType = Crypto::SpongeType::KERL);
   virtual ~Extended();
